Moved shadow related code from LightweightPipeline.cs into a proper shadow pass class
