+++ ScriptableRenderPipeline/LightweightPipeline/LWRP/LightweightShadowPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 51015cd4dddd59d4b95f01cf645067bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
